split database_interop into multiple files
