+++ packages/firebase_database/firebase_database_web/lib/src/interop/database_interop.dart
@@ -13,6 +13,10 @@ import 'package:firebase_core_web/firebase_core_web_interop.dart'
 import 'package:firebase_database_web/src/interop/app_interop.dart';
 import 'package:js/js.dart';
 
+part 'query_interop.dart';
+part 'reference_interop.dart';
+part 'data_snapshot_interop.dart';
+
 external void enableLogging([logger, bool persistent]);
 
 // ignore: avoid_classes_with_only_static_members
@@ -39,143 +43,6 @@ abstract class DatabaseJsImpl {
   external ReferenceJsImpl ref([String? path]);
 
   external ReferenceJsImpl refFromURL(String url);
-}
-
-@JS('Reference')
-abstract class ReferenceJsImpl extends QueryJsImpl {
-  external String get key;
-
-  external set key(String s);
-
-  external ReferenceJsImpl get parent;
-
-  external set parent(ReferenceJsImpl r);
-
-  external ReferenceJsImpl get root;
-
-  external set root(ReferenceJsImpl r);
-
-  external ReferenceJsImpl child(String path);
-
-  external OnDisconnectJsImpl onDisconnect();
-
-  external ThenableReferenceJsImpl push([
-    value,
-    void Function(dynamic) onComplete,
-  ]);
-
-  external core_interop.PromiseJsImpl<void> remove(
-      [void Function(dynamic) onComplete]);
-
-  external core_interop.PromiseJsImpl<void> set(value,
-      [void Function(dynamic) onComplete]);
-
-  external core_interop.PromiseJsImpl<void> setPriority(
-    priority, [
-    void Function(dynamic) onComplete,
-  ]);
-
-  external core_interop.PromiseJsImpl<void> setWithPriority(
-    newVal,
-    newPriority, [
-    void Function(dynamic) onComplete,
-  ]);
-
-  external core_interop.PromiseJsImpl<TransactionJsImpl> transaction(
-    void Function(dynamic) transactionUpdate, [
-    dynamic Function(Object, bool, DataSnapshotJsImpl) onComplete,
-    bool applyLocally,
-  ]);
-
-  external core_interop.PromiseJsImpl<void> update(
-    values, [
-    void Function(dynamic) onComplete,
-  ]);
-}
-
-@JS('Query')
-abstract class QueryJsImpl {
-  external ReferenceJsImpl get ref;
-
-  external set ref(ReferenceJsImpl r);
-
-  external QueryJsImpl endAt(value, [String key]);
-
-  external QueryJsImpl equalTo(value, [String key]);
-
-  external bool isEqual(QueryJsImpl other);
-
-  external QueryJsImpl limitToFirst(int limit);
-
-  external QueryJsImpl limitToLast(int limit);
-
-  external void off([
-    String eventType,
-    dynamic Function(DataSnapshotJsImpl, [String]) callback,
-    context,
-  ]);
-
-  external void Function() on(
-    String eventType,
-    dynamic Function(DataSnapshotJsImpl, [String]) callback, [
-    cancelCallbackOrContext,
-    context,
-  ]);
-
-  external core_interop.PromiseJsImpl<dynamic> once(
-    String eventType, [
-    dynamic Function(DataSnapshotJsImpl, [String]) callback,
-    successCallback,
-    failureCallbackOrContext,
-    context,
-  ]);
-
-  external QueryJsImpl orderByChild(String path);
-
-  external QueryJsImpl orderByKey();
-
-  external QueryJsImpl orderByPriority();
-
-  external QueryJsImpl orderByValue();
-
-  external QueryJsImpl startAt(value, [String key]);
-
-  external Object toJSON();
-
-  @override
-  external String toString();
-}
-
-@JS('DataSnapshot')
-@anonymous
-abstract class DataSnapshotJsImpl {
-  external String get key;
-
-  external set key(String s);
-
-  external ReferenceJsImpl get ref;
-
-  external set ref(ReferenceJsImpl r);
-
-  external DataSnapshotJsImpl child(String path);
-
-  external bool exists();
-
-  external dynamic exportVal();
-
-  external bool forEach(void Function(dynamic) action);
-
-  external dynamic getPriority();
-
-  external bool hasChild(String path);
-
-  external bool hasChildren();
-
-  external int numChildren();
-
-  external dynamic val();
-
-  external Object toJSON();
 }
 
 @JS('OnDisconnect')
