+++ packages/firebase_database/firebase_database_web/lib/src/interop/query_interop.dart
@@ -1,0 +1,54 @@
+part of firebase.database_interop;
+
+@JS('Query')
+abstract class QueryJsImpl {
+  external ReferenceJsImpl get ref;
+
+  external set ref(ReferenceJsImpl r);
+
+  external QueryJsImpl endAt(value, [String key]);
+
+  external QueryJsImpl equalTo(value, [String key]);
+
+  external bool isEqual(QueryJsImpl other);
+
+  external QueryJsImpl limitToFirst(int limit);
+
+  external QueryJsImpl limitToLast(int limit);
+
+  external void off([
+    String eventType,
+    dynamic Function(DataSnapshotJsImpl, [String]) callback,
+    context,
+  ]);
+
+  external void Function() on(
+    String eventType,
+    dynamic Function(DataSnapshotJsImpl, [String]) callback, [
+    cancelCallbackOrContext,
+    context,
+  ]);
+
+  external core_interop.PromiseJsImpl<dynamic> once(
+    String eventType, [
+    dynamic Function(DataSnapshotJsImpl, [String]) callback,
+    successCallback,
+    failureCallbackOrContext,
+    context,
+  ]);
+
+  external QueryJsImpl orderByChild(String path);
+
+  external QueryJsImpl orderByKey();
+
+  external QueryJsImpl orderByPriority();
+
+  external QueryJsImpl orderByValue();
+
+  external QueryJsImpl startAt(value, [String key]);
+
+  external Object toJSON();
+
+  @override
+  external String toString();
+}
